fix: test hanging error fix
