+++ node/pkg/admin/tests/main_test.go
@@ -55,6 +55,7 @@ func cleanup() {
 func TestMain(m *testing.M) {
 	// setup
 	code := m.Run()
+
 	db.ClosePool()
 	db.CloseRedis()
 	// teardown
